staging: ft1000: Use specific error return code for reg_ft1000_netdev().

+++ drivers/staging/ft1000/ft1000-usb/ft1000_usb.h
@@ -597,7 +597,7 @@ char *getfw (char *fn, size_t *pimgsz);
 int dsp_reload(struct ft1000_device *ft1000dev);
 u16 init_ft1000_netdev(struct ft1000_device *ft1000dev);
 struct usb_interface;
-u16 reg_ft1000_netdev(struct ft1000_device *ft1000dev, struct usb_interface *intf);
+int reg_ft1000_netdev(struct ft1000_device *ft1000dev, struct usb_interface *intf);
 int ft1000_poll(void* dev_id);
 
 void ft1000InitProc(struct net_device *dev);
